a function that prints all the elements of a listint_t list

+++ 0x13-more_singly_linked_lists/0-print_listint.c
@@ -8,11 +8,11 @@
  */
 size_t print_listint(const listint_t *h)
 {
-	int a = 0;
+	size_t a = 0;
 
 	while (h != NULL)
 	{
-		printf("%i\n", h->n);
+		printf("%d\n", h->n);
 		h = h->next;
 		a++;
 	}
